Additional istructions for adding the config via UI when installed with HACS
Thanks for creating this!

I installed it via HACS today and tried to paste your example as a manual card, but it didn't work and I had to play around a little until I managed to get it working.
Updating the Readme to make it more easy for future users that are not using `ui-lovelace.yaml`. 

Please edit to your likings, wasn't sure how to sub categorize the config sections.